nice

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: nyx <nyx@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 14:49:11 by nyx               #+#    #+#             */
-/*   Updated: 2024/11/04 14:49:12 by nyx              ###   ########.fr       */
+/*   Updated: 2024/11/04 21:46:43 by nyx              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	size_t i;
 	i = 0;
