+++ ft_strmapi.c
@@ -6,15 +6,16 @@
 /*   By: nyx <nyx@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 14:49:14 by nyx               #+#    #+#             */
-/*   Updated: 2024/11/04 14:49:15 by nyx              ###   ########.fr       */
+/*   Updated: 2024/11/04 21:45:32 by nyx              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdlib.h>
 
 char	*ft_strmapi(const char *s, char (*f)(unsigned int, char))
 {
-	unsigned int i;
+	size_t i;
 	char *res;
 
 	res = malloc((ft_strlen(s) + 1) * sizeof(char));
